Refactor tests in detekt-gradle-plugin
1. Moved data classes, DslGradleRunner, and DslTestBuilder under `testkit`
2. Updated because reason in build.gradle.kts
3. Inlined and updated the formatting version under test.
