+++ detekt-gradle-plugin/build.gradle.kts
@@ -18,7 +18,7 @@ dependencies {
 
     constraints {
         implementation("org.jetbrains.kotlin:kotlin-reflect:1.4.0") {
-            because("com.android.tools.build:gradle:4.0.1 depends on Kotlin 1.3.72 and we should not mix 1.3 and 1.4")
+            because("$androidGradlePlugin depends on Kotlin 1.3.72 and we should not mix 1.3 and 1.4")
         }
     }
 }
