--- detekt-gradle-plugin/src/test/kotlin/io/gitlab/arturbosch/detekt/TestVersion.kt
@@ -1,3 +1,0 @@
-package io.gitlab.arturbosch.detekt
-
-const val VERSION_UNDER_TEST: String = "1.0.0.RC9.2"
