+++ detekt-gradle-plugin/src/test/kotlin/io/gitlab/arturbosch/detekt/testkit/ProjectLayout.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt
+package io.gitlab.arturbosch.detekt.testkit
 
 data class ProjectLayout(
     val numberOfSourceFilesInRootPerSourceDir: Int,
